gazelle: move tar.go to the fetch package
