+++ .bazelversion
